pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["chilledsnake <chilledsnake.pl@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.5.2"
"backports.zoneinfo" = "0.2.1"
CacheControl = "0.12.11"
cachy = "0.3.0"
certifi = "2022.5.18.1"
charset-normalizer = "2.0.12"
cleo = "0.8.1"
clikit = "0.6.2"
crashtest = "0.3.1"
distlib = "0.3.4"
Django = "4.0.4"
django-filter = "21.1"
djangorestframework = "3.13.1"
filelock = "3.7.0"
greenlet = "1.1.2"
html5lib = "1.1"
idna = "3.3"
importlib-metadata = "4.11.4"
keyring = "23.5.1"
lockfile = "0.12.2"
Markdown = "3.3.7"
msgpack = "1.0.3"
packaging = "20.9"
pastel = "0.2.1"
pexpect = "4.8.0"
pkginfo = "1.8.2"
platformdirs = "2.5.2"
poetry = "1.1.13"
poetry-core = "1.0.8"
psycopg2-binary = "2.9.3"
ptyprocess = "0.7.0"
pylev = "1.4.0"
pyparsing = "3.0.9"
pytz = "2022.1"
requests = "2.27.1"
requests-toolbelt = "0.9.1"
shellingham = "1.4.0"
six = "1.16.0"
sqlparse = "0.4.2"
tomlkit = "0.11.0"
urllib3 = "1.26.9"
virtualenv = "20.14.1"
webencodings = "0.5.1"
zipp = "3.8.0"
"zope.event" = "4.5.0"
"zope.interface" = "5.4.0"
python-decouple = "^3.6"
dj-database-url = "^0.5.0"
gunicorn = "^20.1.0"
whitenoise = "^6.1.0"
black = "^22.3.0"
isort = "^5.10.1"
drf-yasg = "^1.20.0"
django-storages = "^1.13.1"
boto3 = "^1.24.77"
Pillow = "^9.2.0"
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
